refactor: consolidate exhibit data into constants.ts
