+++ app/components/common/cards/small/exhibit/index.tsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import { Icon } from '@trussworks/react-uswds';
 import SmallCard from '../index';
-import type { Exhibit } from 'app/config';
+import type { Exhibit } from 'app/types';
 
 interface FormattedSectionProps {
   heading: string;
